图片批处理脚本/requirements.txt
openai>=1.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
Pillow>=10.2.0
tqdm>=4.66.0
requests>=2.31.0
